--- src/use_case/quiz/display_quiz/DisplayQuizOutputBoundary.java
@@ -1,6 +1,0 @@
-package use_case.quiz.display_quiz;
-
-public interface DisplayQuizOutputBoundary {
-    void prepareSuccessView(DisplayQuizOutputData quiz);
-    void prepareFailView(String error);
-}
