+++ src/use_case/quiz/display_quiz/DisplayQuizzesInputBoundary.java
@@ -1,5 +1,5 @@
 package use_case.quiz.display_quiz;
 
-public interface DisplayQuizInputBoundary {
+public interface DisplayQuizzesInputBoundary {
     void execute();
 }
