+++ src/use_case/quiz/display_quiz/DisplayQuizzesInputData.java
@@ -1,9 +1,10 @@
 package use_case.quiz.display_quiz;
 
-public class DisplayQuizInputData{
+public class DisplayQuizzesInputData {
+    //TODO: this class is not needed, we can delete it later
 //    private String title;
 //
-//    public DisplayQuizInputData(String title){
+//    public DisplayQuizzesInputData(String title){
 //        this.title = title;
 //    }
 //
