+++ src/use_case/quiz/take_quiz/TakeQuizOutputBoundary.java
@@ -1,7 +1,5 @@
 package use_case.quiz.take_quiz;
 
-import use_case.quiz.display_quiz.DisplayQuizOutputData;
-
 public interface TakeQuizOutputBoundary {
 
     void prepareSuccessView(TakeQuizOutputData takeQuizOutputData);
